+++ refena/lib/src/action/refresh_action.dart
@@ -1,0 +1,76 @@
+import 'package:refena/src/action/redux_action.dart';
+import 'package:refena/src/async_value.dart';
+import 'package:refena/src/notifier/base_notifier.dart';
+
+/// An action that refreshes the state of a [ReduxNotifier].
+/// The state must be an [AsyncValue] with the data type [T].
+///
+/// This action does the boilerplate process of
+/// - setting the state to [AsyncValue.loading]
+/// - catching errors and setting the state to [AsyncValue.withError]
+/// - setting the state to [AsyncValue.withData] with the result of [refresh]
+///
+/// Example:
+/// class MyRefreshAction extends RefreshAction<MyNotifier, int> {
+///   @override
+///   Future<int> refresh() async {
+///     await Future.delayed(Duration(seconds: 1));
+///     return 42;
+///   }
+/// }
+abstract class RefreshAction<N extends BaseReduxNotifier<AsyncValue<T>>, T>
+    extends AsyncReduxAction<N, AsyncValue<T>> {
+  /// Implement the refresh logic here.
+  Future<T> refresh();
+
+  @override
+  Future<AsyncValue<T>> reduce() async {
+    dispatch(RefreshSetLoadingAction<N, T>());
+    try {
+      return AsyncValue<T>.withData(await refresh());
+    } catch (error, stackTrace) {
+      dispatch(RefreshSetErrorAction<N, T>(error, stackTrace));
+      rethrow;
+    }
+  }
+}
+
+/// Sets the state of a [ReduxNotifier] to [AsyncValue.loading].
+class RefreshSetLoadingAction<N extends BaseReduxNotifier<AsyncValue<T>>, T>
+    extends ReduxAction<N, AsyncValue<T>> {
+  @override
+  AsyncValue<T> reduce() {
+    return AsyncValue<T>.loading();
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return other is RefreshSetLoadingAction;
+  }
+
+  @override
+  int get hashCode => 0;
+}
+
+/// Sets the state of a [ReduxNotifier] to [AsyncValue.withError].
+class RefreshSetErrorAction<N extends BaseReduxNotifier<AsyncValue<T>>, T>
+    extends ReduxAction<N, AsyncValue<T>> {
+  final Object error;
+  final StackTrace stackTrace;
+
+  RefreshSetErrorAction(this.error, this.stackTrace);
+
+  @override
+  AsyncValue<T> reduce() {
+    return AsyncValue<T>.withError(error, stackTrace);
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return other is RefreshSetErrorAction &&
+        other.error == error;
+  }
+
+  @override
+  int get hashCode => error.hashCode;
+}